Shrinks the following data structures: Bitboard SquareDistance from 16k to 4k Pawns::Entry size from 120 bytes to 88 bytes Pawns::HashTable from 1.9MB to 1.4MB Search Reductions from 64k to 16k
Each has also been verified to produce a small speedup.
The total speedup is 2.5%.
No functional change.

bench: 5608839

+++ src/bitboard.h
@@ -59,7 +59,7 @@ const Bitboard Rank6BB = Rank1BB << (8 * 5);
 const Bitboard Rank7BB = Rank1BB << (8 * 6);
 const Bitboard Rank8BB = Rank1BB << (8 * 7);
 
-extern int SquareDistance[SQUARE_NB][SQUARE_NB];
+extern int8_t SquareDistance[SQUARE_NB][SQUARE_NB];
 
 extern Bitboard SquareBB[SQUARE_NB];
 extern Bitboard FileBB[FILE_NB];
